Move getFallbackValue to @utils/data/default-values
