+++ apps/astro-app/src/utils/data.ts
@@ -62,46 +62,6 @@ export function handleRecordOn(
   });
 }
 
-export function getFallbackValue(datatype: "int" | "integer"): number;
-export function getFallbackValue(datatype: "float" | "number"): number;
-export function getFallbackValue(
-  datatype: "string" | "str" | "text" | "enum",
-): string;
-export function getFallbackValue(datatype: "bool" | "boolean"): boolean;
-export function getFallbackValue(
-  datatype: "date" | "time" | "timestamp",
-): string;
-export function getFallbackValue(
-  datatype: Datatype,
-): number | string | boolean | Date | null | undefined;
-export function getFallbackValue(
-  datatype: Datatype,
-): number | string | boolean | Date | null | undefined {
-  switch (datatype) {
-    case "int":
-    case "integer":
-    case "float":
-    case "number":
-      return 0;
-    case "string":
-    case "str":
-    case "text":
-    case "enum":
-      return "";
-    case "bool":
-    case "boolean":
-      return false;
-    case "date":
-      return "0001-01-01";
-    case "time":
-      return "01:00:00";
-    case "timestamp":
-      return new Date().toString();
-    default:
-      return null;
-  }
-}
-
 type formatDatabaseValueParams =
   | { datatype: "int" | "integer" | "float" | "number"; value: number }
   | { datatype: "string" | "str" | "text"; value: string }
